change view implementation

+++ src/main/java/com/podts/rpg/server/model/entity/PlayerEntity.java
@@ -6,7 +6,7 @@ import com.podts.rpg.server.model.universe.Location;
 
 public class PlayerEntity extends LivingHumanoid {
 	
-	public static double DEFAULT_VIEW_DISTANCE = 15d;
+	public static int DEFAULT_VIEW_DISTANCE = 15;
 	
 	private final Player player;
